move constants out of single package file
