+++ custom/src/test/java/com/splunk/opentelemetry/appd/AppdBonusSpanProcessorTest.java
@@ -16,11 +16,11 @@
 
 package com.splunk.opentelemetry.appd;
 
-import static com.splunk.opentelemetry.appd.AppdBonusConstants.APPD_ATTR_ACCT;
-import static com.splunk.opentelemetry.appd.AppdBonusConstants.APPD_ATTR_APP;
-import static com.splunk.opentelemetry.appd.AppdBonusConstants.APPD_ATTR_BT;
-import static com.splunk.opentelemetry.appd.AppdBonusConstants.APPD_ATTR_TIER;
 import static com.splunk.opentelemetry.appd.AppdBonusPropagator.CONTEXT_KEY;
+import static com.splunk.opentelemetry.appd.AppdBonusSpanProcessor.APPD_ATTR_ACCT;
+import static com.splunk.opentelemetry.appd.AppdBonusSpanProcessor.APPD_ATTR_APP;
+import static com.splunk.opentelemetry.appd.AppdBonusSpanProcessor.APPD_ATTR_BT;
+import static com.splunk.opentelemetry.appd.AppdBonusSpanProcessor.APPD_ATTR_TIER;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
